requirements.txt
pyyaml==6.0.2
aiohttp==3.11.14

# Development dependencies
black==25.1.0
isort==6.0.1
pytest==8.3.4